backend/requirements.txt
python-dotenv
crewai
requests